n-grok-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.135.2"}
psycopg = {extras = ["binary"], version = "^3.3.3"}
httpx = "^0.28.1"
python-dotenv = "^1.2.2"
python-multipart = "^0.0.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
